Convert currency for minimum spend and max discount (voucher)

+++ src/main/java/com/traveloveapi/DTO/voucher/VoucherDTO.java
@@ -10,10 +10,10 @@ import lombok.Data;
 public class VoucherDTO {
     private String voucher_id;
     private VoucherDiscountType type;
-    private float minimum_spend;
+    private Double minimum_spend;
     private Double fixed_discount;
     private float percent_discount;
-    private float max_discount;
+    private Double max_discount;
     private String originCurrency;
     private VoucherTargetType target_type;
     private String target_id;
